Crate framework

Modules

natives

Macros

pop_vec_arg

Used to pop a Vec<Vec> argument off the stack.

Structs

BuildOptions

Represents a set of options for building artifacts from Move.

BuiltPackage

Represents a built package. It allows to extract PackageMetadata. Can also be used to just build Move code.

ReleaseBundle

A release bundle consists of a list of release packages.

ReleaseOptions

Options to configure the generation of a release.

ReleasePackage

A release package consists of package metdata and the code.

RuntimeModuleMetadata

Aptos specific metadata attached to the metadata section of file_format.

Enums

ReleaseTarget

Represents the available release targets. Current is in sync with the current client branch, which is ensured by tests.

Constants

METADATA_FILE_NAME

RELEASE_BUNDLE_EXTENSION

UPGRADE_POLICY_CUSTOM_FIELD

Statics

APTOS_METADATA_KEY

The key used to identify the metadata in the metadata section of the module bytecode. This is more or less arbitrary, besides we should use some unique key to identify Aptos specific metadata (aptos:: here).

Functions

named_addresses

path_in_crate

testnet_release_bundle

Returns the release bundle with which the last testnet was build or updated.

unzip_metadata

unzip_metadata_str

zip_metadata

zip_metadata_str