Module framework::natives::code

Structs

GasParameters

ModuleMetadata

MoveOption

A wrapper around the representation of a Move Option, which is a vector with 0 or 1 element. TODO: move this elsewhere for reuse?

NativeCodeContext

The native code context.

PackageDep

PackageMetadata

The PackageMetadata type. This must be kept in sync with code.move. Documentation is also found there.

PackageRegistry

The package registry at the given address.

PublishRequest

Represents a request for code publishing made from a native call and to be processed by the Aptos VM.

RequestPublishGasParameters

UpgradePolicy

Functions

make_all

make_native_request_publish