Module framework::natives::cryptography::ristretto255

Structs

GasParameters

Functions

make_all

pop_32_byte_slice

Pops a 32 byte slice off the argument stack.

pop_64_byte_slice

Pops a 64 byte slice off the argument stack.

pop_scalar_from_bytes

Pops a Scalar off the argument stack when the argument was a vector.

scalar_from_struct

Get a curve25519-dalek Scalar struct from a Move Scalar struct.

scalar_from_valid_bytes

Constructs a curve25519-dalek Scalar from a sequence of bytes which are assumed to canonically-encode it. Callers who are not sure of the canonicity of the encoding MUST call Scalar::is_canonical() after on the returned Scalar.