Struct framework::natives::transaction_context::NativeTransactionContext

pub struct NativeTransactionContext { /* private fields */ }

The native transaction context extension. This needs to be attached to the NativeContextExtensions value which is passed into session functions, so its accessible from natives of this extension.

Implementations

impl NativeTransactionContext

pub fn new(script_hash: Vec<u8>) -> Self

Create a new instance of a native transaction context. This must be passed in via an extension into VM session functions.