Struct framework::BuiltPackage

pub struct BuiltPackage { /* private fields */ }

Represents a built package. It allows to extract PackageMetadata. Can also be used to just build Move code.

Implementations

impl BuiltPackage

pub fn build(package_path: PathBuf, options: BuildOptions) -> Result<Self>

Builds the package and on success delivers a BuiltPackage.

This function currently reports all Move compilation errors and warnings to stdout, and is not Ok if there was an error among those.

pub fn name(&self) -> &str

Returns the name of this package.

pub fn package_path(&self) -> &Path

pub fn package_artifacts_path(&self) -> PathBuf

pub fn extract_code(&self) -> Vec<Vec<u8>>

Extracts the bytecode for the modules of the built package.

pub fn extract_abis(&self) -> Option<Vec<EntryABI>>

Returns the abis for this package, if available.

pub fn modules(&self) -> impl Iterator<Item = &CompiledModule>

Returns an iterator for all compiled proper (non-script) modules.

pub fn script_count(&self) -> usize

Returns the number of scripts in the package.

pub fn extract_script_code(&self) -> Vec<Vec<u8>>

Returns the serialized bytecode of the scripts in the package.

pub fn extract_metadata(&self) -> Result<PackageMetadata>

Extracts metadata, as needed for releasing a package, from the built package.

pub fn extract_metadata_and_save(&self) -> Result<()>